+++ src/main/java/command/music/RecommendationCommand.java
@@ -1,4 +1,4 @@
-package command;
+package command.music;
 
 import YouTubeSearchApi.YoutubeClient;
 import YouTubeSearchApi.entity.YoutubePlaylist;
