+++ src/main/java/command/music/ShuffleCommand.java
@@ -1,24 +1,31 @@
-package command;
+package command.music;
 
 import client.NanoClient;
 import com.jagrosh.jdautilities.command.Command;
 import com.jagrosh.jdautilities.command.CommandEvent;
+import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import net.dv8tion.jda.api.entities.VoiceChannel;
 import service.music.GuildMusicManager;
 import service.music.HelpProcess;
 import service.music.MusicUtils;
 
-public class PauseCommand extends Command {
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.concurrent.BlockingQueue;
+
+public class ShuffleCommand extends Command {
 
     NanoClient nanoClient;
 
-    public PauseCommand(NanoClient nanoClient) {
+    public ShuffleCommand(NanoClient nanoClient) {
         this.nanoClient = nanoClient;
 
-        this.name = "pause";
+        this.name = "shuffle";
+        this.aliases = new String[] {"shuffle_queue"};
+        this.help = "Shuffles song queue";
         this.guildOnly = true;
-        this.cooldown = 1;
-        this.help = "Pause current playing song";
+        this.cooldown = 2;
         this.category = new Category("Music");
         this.help = HelpProcess.getHelp(this);
     }
@@ -32,9 +39,7 @@ public class PauseCommand extends Command {
         }
 
         GuildMusicManager musicManager = this.nanoClient.getGuildAudioPlayer(event.getGuild());
-
-        if (musicManager.player.getPlayingTrack() == null) {
-            event.reply(":x: | Not currently playing anything.");
+        if (musicManager.scheduler.getQueue().size() <= 0) {
             return;
         }
 
@@ -45,14 +50,19 @@ public class PauseCommand extends Command {
             }
         }
 
-        if (!musicManager.isPauseStatus()) {
-            musicManager.player.setPaused(true);
-            musicManager.setPauseStatus(true);
-        }
-        else {
-            musicManager.player.setPaused(false);
-            musicManager.setPauseStatus(false);
-        }
-        event.getMessage().addReaction("\u23F8").queue(); // Pause button
+        // Get current queue
+        BlockingQueue<AudioTrack> queue = musicManager.scheduler.getQueue();
+
+        // Add songs to List
+        List<AudioTrack> tracks = new ArrayList<>();
+        tracks.addAll(queue);
+
+        Collections.shuffle(tracks);
+
+        // Replace current queue
+        musicManager.scheduler.getQueue().clear();
+        musicManager.scheduler.getQueue().addAll(tracks);
+
+        event.getMessage().addReaction(event.getClient().getSuccess()).queue();
     }
 }
