+++ src/main/java/command/music/YoutubeSearchCommand.java
@@ -1,4 +1,4 @@
-package command;
+package command.music;
 
 import YouTubeSearchApi.*;
 import YouTubeSearchApi.entity.YoutubeVideo;
